+++ class/scrumkanban.class.php
@@ -27,6 +27,7 @@ require_once DOL_DOCUMENT_ROOT.'/core/class/commonobject.class.php';
 require_once __DIR__ . '/scrumkanbanlist.class.php';
 require_once __DIR__ . '/scrumcard.class.php';
 require_once __DIR__ . '/../lib/scrumproject.lib.php';
+require_once __DIR__ . '/commonObjectQuickTools.trait.php';
 
 //require_once DOL_DOCUMENT_ROOT . '/societe/class/societe.class.php';
 //require_once DOL_DOCUMENT_ROOT . '/product/class/product.class.php';
@@ -36,6 +37,8 @@ require_once __DIR__ . '/../lib/scrumproject.lib.php';
  */
 class ScrumKanban extends CommonObject
 {
+
+	use CommonObjectQuickTools;
 
 //	/**
 //	 * @var string ID of module.