+++ ChangeLog.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- NEW : Visual color display for project planning *23/01/2022* - 1.32.0
 - NEW : Unspash background search for kanban *20/01/2022* - 1.31.0
 - NEW : Kanban background selection *20/01/2022* - 1.30.0  
   Note : add new Rights for kanban too